# web/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swarm Builder",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/graphs/{graph_id}/runs": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Start Run",
        "description": "Start one run and return its job id.\n\nThe input is coerced up front so a bad input is a 422 here rather\nthan a failed job. Staleness is checked here too, so the response can\nsay whether a compile will precede the run.",
        "operationId": "start_run_api_graphs__graph_id__runs_post",
        "parameters": [
          {
            "name": "graph_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Graph Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartRunResponse"
                }
              }
            }
          },
          "404": {
            "description": "no such graph"
          },
          "409": {
            "description": "a compile or run is already live, or the project is stale"
          },
          "422": {
            "description": "the input does not fit the entry node's port type"
          },
          "503": {
            "description": "the run subsystem is unavailable"
          }
        }
      },
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "List Runs Route",
        "description": "Persisted runs for a graph, newest first (at most 20 are kept).",
        "operationId": "list_runs_route_api_graphs__graph_id__runs_get",
        "parameters": [
          {
            "name": "graph_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Graph Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/graphs/{graph_id}/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Get Run Route",
        "description": "One persisted run record (the live view is the job snapshot).",
        "operationId": "get_run_route_api_graphs__graph_id__runs__run_id__get",
        "parameters": [
          {
            "name": "graph_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Graph Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "404": {
            "description": "no such run record"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeRunRecord": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "inputs": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Inputs"
          },
          "output": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "stateDelta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statedelta"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "durationMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "NodeRunRecord",
        "description": "What one step reported during a run (the last ``node`` frame wins)."
      },
...
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunRecord"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "runs"
        ],
        "title": "RunListResponse"
      },
      "RunRecord": {
        "properties": {
          "runId": {
            "type": "string",
            "title": "Runid"
          },
          "graphId": {
            "type": "string",
            "title": "Graphid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "input": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "output": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "state": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "durationMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationms"
          },
          "compiled": {
            "type": "boolean",
            "title": "Compiled",
            "default": false
          },
          "dryRun": {
            "type": "boolean",
            "title": "Dryrun",
            "default": false
          },
          "nodes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeRunRecord"
            },
            "type": "object",
            "title": "Nodes"
          }
        },
        "type": "object",
        "required": [
          "runId",
          "graphId",
          "status",
          "createdAt"
        ],
        "title": "RunRecord",
        "description": "One persisted run."
      },
...
      "StartRunRequest": {
        "properties": {
          "input": {
            "title": "Input"
          },
          "compileIfStale": {
            "type": "boolean",
            "title": "Compileifstale",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "StartRunRequest",
        "description": "Body of ``POST /api/graphs/:id/runs``.\n\n``input`` is interpreted by the entry node's ``inputType``\n(``compile/run.py``'s ``coerce_input``): a ``str`` port takes the\nstring verbatim; ``json``/``list[str]`` take the parsed value or a\nJSON string of it. ``compileIfStale`` (default on) recompiles a\nmissing or out-of-date project before running; off, a stale project\nis a 409."
      },
      "StartRunResponse": {
        "properties": {
          "runId": {
            "type": "string",
            "title": "Runid"
          },
          "willCompile": {
            "type": "boolean",
            "title": "Willcompile"
          },
          "dryRun": {
            "type": "boolean",
            "title": "Dryrun"
          }
        },
        "type": "object",
        "required": [
          "runId",
          "willCompile",
          "dryRun"
        ],
        "title": "StartRunResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}